Move other C functions to module c::misc
Reimplement with module c::common.
Format with rustfmt.
